scripts/requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
cryptography>=42.0.0
pytest>=8.0.0
requests>=2.31.0
